server/requirements.txt
flask
kubernetes
pyyaml